requirements.txt
Django==5.0.6
django-ninja==1.1.0
python-dotenv==1.0.1
